requirements.txt
asgiref==3.8.1
cssselect==1.3.0
d==0.2.2
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
lxml==5.3.2
Markdown==3.8
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pyquery==2.0.1
sqlparse==0.5.3
django-cors-headers==4.7.0